+++ R/ICESrectangles-datadoc.R
@@ -26,7 +26,7 @@
 #'
 #' @usage data(ICESrectangles)
 #'
-#' @format \code{\link[sp]{SpatialPolygonsDataFrame}} with rectange names identified in the column 'StratumName'. See \code{\link[RstoxBase]{StratumPolygon}}.
+#' @format \code{\link[sf]{sf}} with area names identified in the column 'StratumName'. See \code{\link[RstoxBase]{StratumPolygon}} (v.2).
 #'
 #' @concept area code polygons
 #' @keywords datasets