+++ R/NAFOareas-datadoc.R
@@ -16,7 +16,7 @@
 #'
 #' @usage data(NAFOareas)
 #'
-#' @format \code{\link[sp]{SpatialPolygonsDataFrame}} with location names identified in the column 'StratumName'. See \code{\link[RstoxBase]{StratumPolygon}}.
+#' @format \code{\link[sf]{sf}} with area names identified in the column 'StratumName'. See \code{\link[RstoxBase]{StratumPolygon}} (v.2).
 #'
 #' @concept area code polygons
 #' @keywords datasets
@@ -28,5 +28,5 @@
 #'  RstoxFDA::plotArea(areaDef = combo)
 #'
 #'  # conversion table Norwegian and international convention
-#'  RstoxFDA::NAFOareas@data[,c("homr", "nafo_names")]
+#'  sf::st_drop_geometry(RstoxFDA::NAFOareas[,c("homr", "nafo_names")])
 "NAFOareas"