+++ inst/tinytest/test-StoxBaselineFunctions.R
@@ -689,7 +689,6 @@ expect_error(RstoxFDA::AddAreaPositionStoxLanding(stoxLandingPre, areaPos), "Col
 #context("test-StoxBaselineFunctions: AppendStratumStoxLanding")
 
 strp <- RstoxFDA::mainareaFdir2018
-#sp::proj4string(strp) <- sp::CRS("+proj=longlat +datum=WGS84")
 
 areafile <- system.file("testresources","mainarea_fdir_from_2018_compl.txt", package="RstoxFDA")
 areaPos <- RstoxFDA::DefineAreaPosition(NULL, FileName = areafile, StratumPolygon = NULL)